# src/main/java/com/sonal/sportsbetting/dto/SettleEventResponse.java
package com.sonal.sportsbetting.dto;

import java.math.BigDecimal;

public record SettleEventResponse(
        String eventId,
        String winningSelection,
        int winners,
        int losers,
        BigDecimal totalPayout,
        BigDecimal globalExposure
) {
}