# src/main/java/com/sonal/sportsbetting/error/ApiErrorResponse.java
package com.sonal.sportsbetting.error;

import java.time.Instant;
import java.util.Map;

public record ApiErrorResponse(
        String code,
        String message,
        Map<String, String> fieldErrors,
        Instant timestamp
) {
}