# src/main/java/com/sonal/sportsbetting/model/OddsUpdate.java
package com.sonal.sportsbetting.model;

import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.NotNull;
import jakarta.validation.constraints.Positive;

import java.math.BigDecimal;

public class OddsUpdate {
    @NotBlank
    private String eventId;
    @NotBlank
    private String selection;
    @NotNull
    @Positive
    private BigDecimal odds;

    public String getEventId() {
        return eventId;
    }

    public void setEventId(String eventId) {
        this.eventId = eventId;
    }

    public String getSelection() {
        return selection;
    }

    public void setSelection(String selection) {
        this.selection = selection;
    }

    public BigDecimal getOdds() {
        return odds;
    }

    public void setOdds(BigDecimal odds) {
        this.odds = odds;
    }
}